添加关于和协议管理接口与实现；增加获取更多设置功能
新增对`about`和`protocol`模块的管理API，包括添加、修改、删除及查询操作。同时，对内部服务进行了相应的扩展以支持这些新功能。此外，迷你端增加了获取关于和协议设置的功能。

在`internal/cmd/cmd.go`中添加了`about`和`protocol`相关的路由处理；
在`hack/config.yaml`中添加了新的数据表配置项；
在`internal/model/do/re_take_list.go`中调整了列定义，将`Id`字段类型从注释中的字符串更改为实际使用的接口类型；
在`internal/model/entity/re_order_list.go`中，将`Balance`字段的数据类型从浮点数更改为整型；
更新了`README.md`文件，加入了构建镜像的指令。

注意：此变更集涉及的`.gitignore`更改可能是一个误操作，因为没有明确的说明为什么新增了`data`和`wechat`这两个条目，也没有解释移除`**/config/config.yaml`的具体原因。建议在提交前进一步澄清这些改动。

+++ internal/model/entity/re_order_list.go
@@ -14,7 +14,7 @@ type ReOrderList struct {
 	UserId     int         `json:"user_id"     orm:"user_id"     ` // 用户ID
 	RecyclerId int         `json:"recycler_id" orm:"recycler_id" ` // 回收员ID
 	Kg         float64     `json:"kg"          orm:"kg"          ` // 公斤
-	Balance    float64     `json:"balance"     orm:"balance"     ` // 红包
+	Balance    int         `json:"balance"     orm:"balance"     ` // 红包
 	Name       string      `json:"name"        orm:"name"        ` // 名字
 	Phone      string      `json:"phone"       orm:"phone"       ` // 手机号
 	Provinces  string      `json:"provinces"   orm:"provinces"   ` // 省
